+++ curso-spring-restapi/src/main/java/com/jpbandeira/springrestapi/domain/Categoria.java
@@ -15,7 +15,7 @@ public class Categoria implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	private String nome;
 	@JsonManagedReference
 	@OneToMany(mappedBy = "categoria", cascade = CascadeType.ALL)
@@ -23,7 +23,7 @@ public class Categoria implements Serializable {
 	
 	public Categoria() {}
 
-	public Categoria(Long id ,String nome) {
+	public Categoria(Integer id ,String nome) {
 		this.id = id;
 		this.nome = nome;
 	}
