+++ curso-spring-restapi/src/main/java/com/jpbandeira/springrestapi/domain/Endereco.java
@@ -12,7 +12,7 @@ import java.io.Serializable;
 public class Endereco  implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String logadouro;
     private String numero;
     private String complemento;
@@ -30,7 +30,7 @@ public class Endereco  implements Serializable {
 
     public Endereco(){}
 
-    public Endereco(Long id ,String logadouro, String numero, String complemento, String bairro, String cep, Cliente cliente, Cidade cidade) {
+    public Endereco(Integer id ,String logadouro, String numero, String complemento, String bairro, String cep, Cliente cliente, Cidade cidade) {
         this.id = id;
         this.logadouro = logadouro;
         this.numero = numero;
