+++ curso-spring-restapi/src/main/java/com/jpbandeira/springrestapi/domain/Estado.java
@@ -16,7 +16,7 @@ import java.util.Objects;
 public class Estado  implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String nome;
     @JsonIgnore
     @OneToMany(mappedBy = "estado", cascade = CascadeType.ALL)
@@ -25,7 +25,7 @@ public class Estado  implements Serializable {
     public Estado() {
     }
 
-    public Estado(Long id ,String nome) {
+    public Estado(Integer id ,String nome) {
         this.id = id;
         this.nome = nome;
     }
