+++ curso-spring-restapi/src/main/java/com/jpbandeira/springrestapi/domain/PagamentoComBoleto.java
@@ -19,7 +19,7 @@ public class PagamentoComBoleto extends Pagamento {
         super();
     }
 
-    public PagamentoComBoleto(Long id ,EstadoPagamento estadoPagamento, Pedido pedido, Date dataVencimento, Date dataPagamento) {
+    public PagamentoComBoleto(Integer id ,EstadoPagamento estadoPagamento, Pedido pedido, Date dataVencimento, Date dataPagamento) {
         super(id ,estadoPagamento, pedido);
         this.dataVencimento = dataVencimento;
         this.dataPagamento = dataPagamento;
