+++ curso-spring-restapi/src/main/java/com/jpbandeira/springrestapi/domain/PagamentoComCartao.java
@@ -16,7 +16,7 @@ public class PagamentoComCartao extends Pagamento {
         super();
     }
 
-    public PagamentoComCartao(Long id ,EstadoPagamento estadoPagamento, Pedido pedido, Integer numeroDeParcelas) {
+    public PagamentoComCartao(Integer id ,EstadoPagamento estadoPagamento, Pedido pedido, Integer numeroDeParcelas) {
         super(id,estadoPagamento, pedido);
         this.numeroDeParcelas = numeroDeParcelas;
     }
