+++ curso-spring-restapi/src/main/java/com/jpbandeira/springrestapi/dto/ClienteNewDto.java
@@ -43,7 +43,7 @@ public class ClienteNewDto implements Serializable {
     private String telefone2;
     private String telefone3;
 
-    private Long cidadeId;
+    private Integer cidadeId;
 
     public ClienteNewDto() {
     }
