+++ curso-spring-restapi/src/main/java/com/jpbandeira/springrestapi/services/CategoriaService.java
@@ -20,7 +20,7 @@ public class CategoriaService {
 	@Autowired
 	private CategoriaRepository repository;
 
-	public Categoria find(Long id) throws ObjectNotFoundException {
+	public Categoria find(Integer id) throws ObjectNotFoundException {
 		Optional<Categoria> objeto = repository.findById(id);
 		return objeto.orElseThrow(() -> new ObjectNotFoundException("Objeto não encontrado! Id: " + id + ", Tipo: " + Categoria.class.getName()));
 	}
@@ -36,7 +36,7 @@ public class CategoriaService {
 		return repository.save(objeto);
 	}
 
-	public void delete(Long id) {
+	public void delete(Integer id) {
 		repository.deleteById(id);
 	}
